Add a data frame to an XML document

This function appends rows from a data frame as child elements to a specified root XML node. Each row is added as a new child element with column values either as attributes or child elements, depending on the structure of the data.

Usage

add_df_to_xml(root, df, row_name)

Arguments

root

An XML node (created with xml_new_root or similar) that will serve as the root element to which the data frame rows will be added.

df

A data frame where each row represents a set of values to be added as a child element to the root XML node. The column names of the data frame are used as the names of the child elements.

row_name

A character string representing the name to be used for each row of the data frame in the XML structure. Each row will be added as a child element under the root node with this name.

Value

The modified root XML node with the added rows from the data frame.

Details

For each row in the data frame, a new child node is created under the root element. Each column in the data frame becomes a child element under that row's node, with the column name used as the tag and the value in the data frame used as the content of that tag. If a value is NA, that column is skipped for that row.

The function is designed to handle simple data frames where each cell contains atomic values (strings, numbers, etc.). It does not handle complex or nested structures.

Examples

if (FALSE) { # \dontrun{
# Example for disease prevalence across age and sex groups
root <- xml2::xml_new_root("prevalences")
df <- data.frame(age = c(21, 21, 22, 22), sex = c(0, 1, 0, 1), value = c(1000, 2000, 1500, 3000))
root <- add_df_to_xml(root, df, "prevalence")
} # }