Collect files on relative risks in compact lists

The function iterates over the keys in keyword_map. For each key, it iterates over the elements in in_list, extracts the element named by the key and binds the dataframes into one, using rbind. It requires that all keys in keyword_map are present as list names the lists of in_list.

Usage

collect_relative_risks(in_list, keyword_map)

Arguments

in_list

A list of lists, where the upper most level are names of risk factors or diseases, and each list contains one or multiple lists corresponding to the names in mapping.

keyword_map

A named list that maps list names on the second nesting of in_list (the keys) to shorter names, or NULL (the values). If the values are not NULL, it is assigned to the to value of the entire list in the output. If values are NULL, the names of the list that are being iterated over are assigned to to in the output.

Value

A list with 2 elements, with the names coming from the names of keyword_map. The two lists contain a list of 3, named from, to, and filename. They indicate the name of the file that contains the relative risks for going from risk factors to diseases, between diseases, or from diseases to death/disability. If for any type of relative risk there is no file present, the list is empty.