Configure a scenario

Creates a configuration list for an intervention scenario in the simulation.

Usage

configure_scenario(
  name,
  transition_filename,
  prevalence_filename,
  success_rate = 100,
  min_age = 0,
  max_age = 95,
  gender = 2
)

Arguments

name

A character string which specifies the unique name of the scenario.

transition_filename

A character string which specifies the path to the transition rates file.

prevalence_filename

A character string which specifies the path to the prevalence file.

success_rate

A numeric which specifies the intervention success rate (0-100). Default is 100.

min_age

A numeric which specifies the minimum target age (0-95). Default is 0.

max_age

A numeric which specifies the maximum target age (0-95). Default is 95.

gender

A numeric which specifies the target gender (0=female, 1=male, 2=both). Default is 2 (both).

Value

A list containing the scenario configuration parameters.

Examples

configure_scenario(
  name = "All_Never_Smokers",
  success_rate = 80,
  min_age = 18,
  max_age = 65,
  gender = 2, # Both female and male
  transition_filename = "NL_RF_smoking_Transitions_Netto",
  prevalence_filename = "RF_smoking_cat_All_Never_Smokers_Prev"
)
#> $uniquename
#> [1] "All_Never_Smokers"
#> 
#> $successRate
#> [1] 80
#> 
#> $targetMinAge
#> [1] 18
#> 
#> $targetMaxAge
#> [1] 65
#> 
#> $targetSex
#> [1] 2
#> 
#> $transfilename
#> [1] "NL_RF_smoking_Transitions_Netto"
#> 
#> $prevfilename
#> [1] "RF_smoking_cat_All_Never_Smokers_Prev"
#>