Download and extract GitHub release asset

Download and extract GitHub release asset

Usage

download_github_release(
  repo_url = "https://github.com/Dynamo-HIA/dynamo-hia",
  release_tag = "v3.0.0-beta.1",
  os = NULL,
  dest_dir = tempdir(),
  make_executable = TRUE,
  github_pat = NULL
)

Arguments

repo_url

Character string of the repository URL. Default is "https://github.com/Dynamo-HIA/dynamo-hia".

release_tag

Character string of the release tag. Default is "v3.0.0-beta.1".

os

Optional character string of the operating system. Default is NULL which tries to detect the OS automatically.

dest_dir

Character string of the destination directory for extraction (default: tempdir()).

make_executable

Logical indicating whether to set execute permissions. Default is TRUE.

github_pat

Optional GitHub Access Token. Default is NULL which will try to use the token from the environment variable GITHUB_PAT.

Value

An invisible fs::path object pointing to the extracted contents.

Examples

if (FALSE) { # \dontrun{
# With default arguments
download_github_release()

# From specific repository and release tag
download_github_release(
  repo_url = "https://github.com/Dynamo-HIA/dynamo-hia",
  release_tag = "v3.0.0-beta.1"
)
} # }