Filter a dataframe based on return values from a list of servers.

Filter a dataframe based on return values from a list of servers.

Usage

filter_df_from_server_data(df_to_filter, server_list, filter_col = "filename")

Arguments

df_to_filter

A data frame to be filtered by the output of server_list.

server_list

A list of servers, each returning a character vector, for instance of user-chosen elements.

filter_col

The column in df_to_filter to use for subsetting relative to user input data.

Value

A data.frame