Filter a list of items

For a list of lists, returns for each element in the list the setdiff to some input.

Usage

filter_items(
  input_item,
  selected_risk_factors,
  risk_factor_feature,
  items_to_filter,
  debug
)

Arguments

input_item

A string with a chosen file name.

selected_risk_factors

A list of two character vectors, named "transitions" and "prevalence". The list is obtained from evaluating the output of risk_factor_server.

risk_factor_feature

A string of either "transitions" or "prevalence".

items_to_filter

A list of lists to be filtered.

debug

A boolean indicating debug mode. If TRUE, the function prints some information messages to the console.

Value

A list of two: keep_items holds the reduced items_to_filter, and default_value holds the value that can be passsed to the selected argument in selectInput. The default_value are defined such that they do not cause an infinite loop if two selectInput fields reactively depend on each other.

Details

The function checks if the input_item is in the selected_risk_factors of the respective risk_factor_feature (prevalences or transitions). If so, it drops the selected risk factor files of the other risk factor feature from the list.