Get disease filenames

Retrieves all filenames associated with a specific disease.

Usage

get_disease_filenames(root_dir, disease_name)

Arguments

root_dir

A character string which specifies the root directory of the project.

disease_name

A character string which specifies the name of the disease directory.

Value

A character vector of filenames associated with the specified disease.

Examples

root_dir <- system.file("extdata", "example-nl", "Reference_Data", package = "hiaR")
get_disease_filenames(root_dir, "Lung_Cancer")
#> /home/runner/work/_temp/Library/hiaR/extdata/example-nl/Reference_Data/Diseases/Lung_Cancer
#> ├── Disability
#> │   └── NL_disease_LungCa_DALY_V1.xml
#> ├── Excess_Mortalities
#> │   └── NL_disease_LungCa_ExecMor_V2.xml
#> ├── Incidences
#> │   └── NL_disease_LungCa_Inc_V2.xml
#> ├── Prevalences
#> │   └── NL_disease_LungCa_Prev_V2.xml
#> └── Relative_Risks_From_Risk_Factor
#>     ├── RR_to_LungCa-Alcohol_cat5.xml
#>     ├── RR_to_LungCa-Smoking_cat3.xml
#>     ├── RR_to_LungCa-Smoking_dur.xml
#>     ├── RR_to_LungCa_V2-BMI_cat3.xml
#>     └── RR_to_LungCa_V2-BMI_cont.xml