hiaR: Dynamic Modeling of Health Impact Assessment in R

The hiaR package provides a user-friendly interface to the DYNAMO-HIA software tool. It contains R functions to configure and run DYNAMO-HIA, and to visualize the results. The package also contains a Shiny app to interactively run DYNAMO-HIA. DYNAMO-HIA (DYNamic MOdeling for Health Impact Assessment) is a generic software tool specially designed to aid the quantification step in health impact assessment. DYNAMO-HIA is generic in the sense that it allows arbitrary chronic diseases or risk-factors that fit the standard causal pathway of the health impact assessment framework to be modeled. Users can import and use their own data to add and analyze diseases or populations of interest.

See also

Author

Maintainer: Hendriek Boshuizen hendriek.boshuizen@rivm.nl

Authors:

Malte Lüken m.luken@esciencecenter.nl

Flavio Hafner f.hafner@esciencecenter.nl

Laurens Bogaardt laurens.bogaardt@rivm.nl